+++ search_query/wos/v_0_0_0/__init__.py
@@ -1,5 +1,0 @@
-#!/usr/bin/env python3
-"""WoS v0.0.0."""
-
-__author__ = """Gerit Wagner"""
-__email__ = "gerit.wagner@hec.ca"